added Twig\Twig as an alias to Twig\Environment
